IBX-11681: Fixed sonar duplicates (#791)
* Fixed sonar duplicates

* Refactor DownloadController tests to use a shared trait for service mocks

---------
