+++ tests/integration/Core/MVC/Symfony/Controller/Content/DownloadControllerRequestFlowTest.php
@@ -8,19 +8,10 @@ declare(strict_types=1);
 
 namespace Ibexa\Tests\Integration\Core\MVC\Symfony\Controller\Content;
 
-use DateTime;
 use Ibexa\Bundle\IO\BinaryStreamResponse;
-use Ibexa\Contracts\Core\Repository\ContentService;
-use Ibexa\Contracts\Core\Repository\Values\Content\ContentInfo;
 use Ibexa\Contracts\Core\Repository\Values\Content\Field;
 use Ibexa\Contracts\Core\Test\IbexaKernelTestCase;
-use Ibexa\Core\FieldType\BinaryFile\Value as BinaryFileValue;
-use Ibexa\Core\Helper\TranslationHelper;
-use Ibexa\Core\IO\IOServiceInterface;
-use Ibexa\Core\IO\Values\BinaryFile;
-use Ibexa\Core\MVC\Symfony\Controller\Content\DownloadController;
-use Ibexa\Core\Repository\Values\Content\Content;
-use Ibexa\Core\Repository\Values\Content\VersionInfo;
+use Ibexa\Tests\Core\MVC\Symfony\Controller\Controller\Content\DownloadControllerTestTrait;
 use Ibexa\Tests\Integration\Core\MVC\Symfony\InternalRoutingTestKernel;
 use Symfony\Component\EventDispatcher\EventDispatcher;
 use Symfony\Component\HttpFoundation\Request;
@@ -44,16 +35,9 @@ use Symfony\Component\Routing\RouteCollection;
  */
 final class DownloadControllerRequestFlowTest extends IbexaKernelTestCase
 {
+    use DownloadControllerTestTrait;
+
     private const FILENAME = 'Q1 report #1 + 100%.jpg';
-
-    /** @var \Ibexa\Contracts\Core\Repository\ContentService&\PHPUnit\Framework\MockObject\MockObject */
-    private ContentService $contentService;
-
-    /** @var \Ibexa\Core\IO\IOServiceInterface&\PHPUnit\Framework\MockObject\MockObject */
-    private IOServiceInterface $ioService;
-
-    /** @var \Ibexa\Core\Helper\TranslationHelper&\PHPUnit\Framework\MockObject\MockObject */
-    private TranslationHelper $translationHelper;
 
     protected static function getKernelClass(): string
     {
@@ -66,9 +50,7 @@ final class DownloadControllerRequestFlowTest extends IbexaKernelTestCase
 
         self::bootKernel();
 
-        $this->contentService = $this->createMock(ContentService::class);
-        $this->ioService = $this->createMock(IOServiceInterface::class);
-        $this->translationHelper = $this->createMock(TranslationHelper::class);
+        $this->initializeServiceMocks();
     }
 
     public function testDownloadsFileWithUrlEncodedFilename(): void
@@ -87,32 +69,15 @@ final class DownloadControllerRequestFlowTest extends IbexaKernelTestCase
 
         self::assertStringContainsString('Q1%20report%20%231%20+%20100%25.jpg', $url);
 
-        $content = $this->createContent();
+        $content = $this->createContent(self::FILENAME);
         $field = $content->getField('file', 'eng-GB');
         self::assertInstanceOf(Field::class, $field);
 
-        $binaryFile = new BinaryFile([
-            'id' => 'binary-file-id',
-            'mtime' => new DateTime(),
-            'size' => 123,
-            'uri' => 'binary-file-uri',
-        ]);
+        $binaryFile = $this->createBinaryFile();
 
-        $this->contentService
-            ->expects(self::once())
-            ->method('loadContent')
-            ->with(42)
-            ->willReturn($content);
-        $this->translationHelper
-            ->expects(self::once())
-            ->method('getTranslatedField')
-            ->with($content, 'file', 'eng-GB')
-            ->willReturn($field);
-        $this->ioService
-            ->expects(self::once())
-            ->method('loadBinaryFile')
-            ->with('binary-file-id')
-            ->willReturn($binaryFile);
+        $this->expectContentLoaded(42, $content);
+        $this->expectTranslatedField($content, 'file', $field);
+        $this->expectBinaryFileLoaded($binaryFile);
 
         $this->configureDownloadController($routes);
 
@@ -131,15 +96,6 @@ final class DownloadControllerRequestFlowTest extends IbexaKernelTestCase
         $route = $routes->get('ibexa.content.download');
         self::assertNotNull($route);
         $route->setDefault('_controller', [$this->createController(), 'downloadBinaryFileAction']);
-    }
-
-    private function createController(): DownloadController
-    {
-        return new DownloadController(
-            $this->contentService,
-            $this->ioService,
-            $this->translationHelper
-        );
     }
 
     private function createHttpKernel(RouteCollection $routes, RequestContext $context): HttpKernel
@@ -173,29 +129,5 @@ final class DownloadControllerRequestFlowTest extends IbexaKernelTestCase
         ));
 
         return $routes;
-    }
-
-    private function createContent(): Content
-    {
-        return new Content([
-            'internalFields' => [
-                new Field([
-                    'fieldDefIdentifier' => 'file',
-                    'languageCode' => 'eng-GB',
-                    'value' => new BinaryFileValue([
-                        'id' => 'binary-file-id',
-                        'fileName' => self::FILENAME,
-                    ]),
-                ]),
-            ],
-            'versionInfo' => new VersionInfo([
-                'contentInfo' => new ContentInfo([
-                    'id' => 42,
-                    'mainLanguageCode' => 'eng-GB',
-                    'name' => 'Test content',
-                    'status' => ContentInfo::STATUS_PUBLISHED,
-                ]),
-            ]),
-        ]);
     }
 }
