+++ tests/lib/MVC/Symfony/Controller/Controller/Content/DownloadControllerTest.php
@@ -8,18 +8,8 @@ declare(strict_types=1);
 
 namespace Ibexa\Tests\Core\MVC\Symfony\Controller\Controller\Content;
 
-use DateTime;
-use Ibexa\Contracts\Core\Repository\ContentService;
-use Ibexa\Contracts\Core\Repository\Values\Content\ContentInfo;
 use Ibexa\Contracts\Core\Repository\Values\Content\Field;
 use Ibexa\Core\Base\Exceptions\NotFoundException as BaseNotFoundException;
-use Ibexa\Core\FieldType\BinaryFile\Value as BinaryFileValue;
-use Ibexa\Core\Helper\TranslationHelper;
-use Ibexa\Core\IO\IOServiceInterface;
-use Ibexa\Core\IO\Values\BinaryFile;
-use Ibexa\Core\MVC\Symfony\Controller\Content\DownloadController;
-use Ibexa\Core\Repository\Values\Content\Content;
-use Ibexa\Core\Repository\Values\Content\VersionInfo;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
@@ -29,81 +19,47 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
  */
 final class DownloadControllerTest extends TestCase
 {
-    /** @var \Ibexa\Contracts\Core\Repository\ContentService&\PHPUnit\Framework\MockObject\MockObject */
-    private ContentService $contentService;
+    use DownloadControllerTestTrait;
 
-    /** @var \Ibexa\Core\IO\IOServiceInterface&\PHPUnit\Framework\MockObject\MockObject */
-    private IOServiceInterface $ioService;
-
-    /** @var \Ibexa\Core\Helper\TranslationHelper&\PHPUnit\Framework\MockObject\MockObject */
-    private TranslationHelper $translationHelper;
+    private const FILENAME = 'Test-file.pdf';
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->contentService = $this->createMock(ContentService::class);
-        $this->ioService = $this->createMock(IOServiceInterface::class);
-        $this->translationHelper = $this->createMock(TranslationHelper::class);
+        $this->initializeServiceMocks();
     }
 
     public function testDownloadBinaryFileActionReturnsBinaryResponseWhenFilenameMatches(): void
     {
-        $content = $this->createContent();
+        $content = $this->createContent(self::FILENAME);
         $field = $content->getField('file', 'eng-GB');
         self::assertInstanceOf(Field::class, $field);
 
         $request = new Request(['inLanguage' => 'eng-GB']);
-        $binaryFile = new BinaryFile([
-            'id' => 'binary-file-id',
-            'mtime' => new DateTime(),
-            'size' => 123,
-            'uri' => 'binary-file-uri',
-        ]);
+        $binaryFile = $this->createBinaryFile();
 
-        $this->contentService
-            ->expects(self::once())
-            ->method('loadContent')
-            ->with(42)
-            ->willReturn($content);
-        $this->translationHelper
-            ->expects(self::once())
-            ->method('getTranslatedField')
-            ->with($content, 'file', 'eng-GB')
-            ->willReturn($field);
-        $this->ioService
-            ->expects(self::once())
-            ->method('loadBinaryFile')
-            ->with('binary-file-id')
-            ->willReturn($binaryFile);
+        $this->expectContentLoaded(42, $content);
+        $this->expectTranslatedField($content, 'file', $field);
+        $this->expectBinaryFileLoaded($binaryFile);
 
-        $response = $this->createController()->downloadBinaryFileAction(42, 'file', 'Test-file.pdf', $request);
+        $response = $this->createController()->downloadBinaryFileAction(42, 'file', self::FILENAME, $request);
 
         self::assertStringContainsString('attachment;', (string) $response->headers->get('Content-Disposition'));
-        self::assertStringContainsString('Test-file.pdf', (string) $response->headers->get('Content-Disposition'));
+        self::assertStringContainsString(self::FILENAME, (string) $response->headers->get('Content-Disposition'));
     }
 
     public function testDownloadBinaryFileActionReturnsNotFoundWhenFilenameDoesNotMatch(): void
     {
-        $content = $this->createContent();
+        $content = $this->createContent(self::FILENAME);
         $field = $content->getField('file', 'eng-GB');
         self::assertInstanceOf(Field::class, $field);
 
         $request = new Request(['inLanguage' => 'eng-GB']);
 
-        $this->contentService
-            ->expects(self::once())
-            ->method('loadContent')
-            ->with(42)
-            ->willReturn($content);
-        $this->translationHelper
-            ->expects(self::once())
-            ->method('getTranslatedField')
-            ->with($content, 'file', 'eng-GB')
-            ->willReturn($field);
-        $this->ioService
-            ->expects($this->never())
-            ->method('loadBinaryFile');
+        $this->expectContentLoaded(42, $content);
+        $this->expectTranslatedField($content, 'file', $field);
+        $this->expectBinaryFileNotLoaded();
 
         $this->assertFileNotFound(function () use ($request): void {
             $this->createController()->downloadBinaryFileAction(42, 'file', 'SomeRandomText.txt', $request);
@@ -112,22 +68,12 @@ final class DownloadControllerTest extends TestCase
 
     public function testDownloadBinaryFileActionReturnsNotFoundWhenFieldIdentifierDoesNotMatch(): void
     {
-        $content = $this->createContent(393, 'New file');
+        $content = $this->createContent(self::FILENAME, 393, 'New file');
         $request = new Request(['inLanguage' => 'eng-GB']);
 
-        $this->contentService
-            ->expects(self::once())
-            ->method('loadContent')
-            ->with(393)
-            ->willReturn($content);
-        $this->translationHelper
-            ->expects(self::once())
-            ->method('getTranslatedField')
-            ->with($content, 'file5', 'eng-GB')
-            ->willReturn(null);
-        $this->ioService
-            ->expects($this->never())
-            ->method('loadBinaryFile');
+        $this->expectContentLoaded(393, $content);
+        $this->expectTranslatedField($content, 'file5', null);
+        $this->expectBinaryFileNotLoaded();
 
         $this->assertFileNotFound(function () use ($request): void {
             $this->createController()->downloadBinaryFileAction(393, 'file5', 'snorelax_snooze.png', $request);
@@ -146,9 +92,7 @@ final class DownloadControllerTest extends TestCase
         $this->translationHelper
             ->expects($this->never())
             ->method('getTranslatedField');
-        $this->ioService
-            ->expects($this->never())
-            ->method('loadBinaryFile');
+        $this->expectBinaryFileNotLoaded();
 
         $this->assertFileNotFound(function () use ($request): void {
             $this->createController()->downloadBinaryFileAction(393, 'file', 'snorelax_snooze.png', $request);
@@ -157,7 +101,7 @@ final class DownloadControllerTest extends TestCase
 
     public function testDownloadBinaryFileByIdActionReturnsNotFoundWhenFieldIdDoesNotMatch(): void
     {
-        $content = $this->createContent();
+        $content = $this->createContent(self::FILENAME);
         $request = new Request();
 
         $this->contentService
@@ -168,22 +112,11 @@ final class DownloadControllerTest extends TestCase
         $this->translationHelper
             ->expects($this->never())
             ->method('getTranslatedField');
-        $this->ioService
-            ->expects($this->never())
-            ->method('loadBinaryFile');
+        $this->expectBinaryFileNotLoaded();
 
         $this->assertFileNotFound(function () use ($request): void {
             $this->createController()->downloadBinaryFileByIdAction($request, 42, 123);
         });
-    }
-
-    private function createController(): DownloadController
-    {
-        return new DownloadController(
-            $this->contentService,
-            $this->ioService,
-            $this->translationHelper
-        );
     }
 
     private function assertFileNotFound(callable $callback): void
@@ -194,30 +127,5 @@ final class DownloadControllerTest extends TestCase
         } catch (NotFoundHttpException $e) {
             self::assertSame('File not found', $e->getMessage());
         }
-    }
-
-    private function createContent(int $contentId = 42, string $contentName = 'Test content'): Content
-    {
-        return new Content([
-            'internalFields' => [
-                new Field([
-                    'id' => 7,
-                    'fieldDefIdentifier' => 'file',
-                    'languageCode' => 'eng-GB',
-                    'value' => new BinaryFileValue([
-                        'id' => 'binary-file-id',
-                        'fileName' => 'Test-file.pdf',
-                    ]),
-                ]),
-            ],
-            'versionInfo' => new VersionInfo([
-                'contentInfo' => new ContentInfo([
-                    'id' => $contentId,
-                    'mainLanguageCode' => 'eng-GB',
-                    'name' => $contentName,
-                    'status' => ContentInfo::STATUS_PUBLISHED,
-                ]),
-            ]),
-        ]);
     }
 }
